Forking and signing upstream 39d58b58f778caed24fd9ebdd8c54ea1e737794c (#1)
Initial checkin: Updated for CBL-C beta

Enabled static linking with CBL (super-experimental)

This only works on macOS so far. It requires you to have a checkout of
`couchbase-lite-c`, and to have built it with CMake, AND you have to
manually postprocess its static libs with `lipo` to work around a Rust
linker bug. :-p

Minor README fixes

Create LICENSE

Co-authored-by: Jens Alfke <jens@couchbase.com>