Setup community edition (#22)
There are two editions of Couchbase Lite C: `enterprise` and
`community`.
In Doctolib we only use the `enterprise` version, but in order to
release this crate on [crates.io](https://crates.io/) it would be
beneficial to offer both options. This is the goal of this pull request.

The main changes of this pull request are:

1. Two new features in Cargo.toml: `enterprise` & `community`. Exactly
one must be selected or the compilation will fail in `build.rs`.
2. All the functionalities that are not available in the `community`
edition (encryption & peer-to-peer) are hidden behind the `enterpise`
feature.
3. Instead of the `libcblite` folder containing the enterprise lib
files, there are now two folders: `libcblite_enterprise` &
`libcblite_community`.
4. The file `build.rs` is modified in order to link with the right
folder depending on the specified edition.
5. The script `update_cblite_c.sh` is enhanced to run the whole process
for both editions.

Some smaller changes are also added:

1. A custom pre-commit hook is added to check that both editions build.
2. The GH workflow `rust` is split in two worflows: `build` & `test`.
3. The three workflows `build`, `test` & `clippy` now use matrixes to
check both editions.
4. The `c_playground` is modified with the new path of the `enterprise`
lib files.