Fix memory leaks & Improve README (#24)
The `Test` Gh workflow is enhanced by also running the whole test suite
with the `LEAK_CHECK` environment variable.
While testing this functionality, it became apparent that we had memory
leaks on `CBLCollection`, `CBLScope` & `CBLQueryIndex` objects.

The functions `retain` and `wrap` were not clear enough on what they
did, especially with the doc comments that were inverted. So the changes
are:

1. Rename `retain` to `reference`
2. Rename `wrap` to `take_ownership`
3. Fix the doc comments on all functions
4. Fix the cases where retain and wrap were misused for `CBLCollection`,
`CBLScope` & `CBLQueryIndex`
5. Implement `Drop` for `QueryIndex`

The tests are now passing with the `LEAK_CHECK` environment variable.

Additional changes:

- The README is improved to document how to test
- The README is improved to make editions (`community` & `enterprise`)
more understandable
- The file `src/main.rs` is deleted as this is a library